refactor: separated weather API logic into custom useWeather hook
